% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intervals-package.R
\docType{package}
\name{intervals}
\alias{intervals}
\alias{intervals-package}
\title{intervals.}
\description{
Package for solving a generaized class of binary learning problems encompassing
hard and soft classification using large-margin classifiers.
}
\author{
Patrick Kimes \email{patrick.kimes@gmail.com}
}

